Make deft--org-follow-link work with archived notes
Try to follow link to deft-directory, if it does not exist tries the
archival directory.  This works best if either deft is not using
recursive-mode at all, or if recursive-archive is on.

Also fixes org-follow-link issue mentioned in
https://github.com/jrblevin/deft/issues/73, and removes (substring
file 1) from deft--org-complete (that was making files without the
leading character in the name)